% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-FSRProportion.R
\name{BestCovBS}
\alias{BestCovBS}
\title{The Selected Covariates Using FSR Variable Selection Method}
\usage{
BestCovBS(DataBSOut, BestLambdaOut)
}
\arguments{
\item{DataBSOut}{The output of \code{\link{DataBS}} function.}

\item{BestLambdaOut}{The output of \code{\link{BestLambdaBS}} function.}
}
\value{
A list consists of two elements
\item{BestER}{A vector of  relevance measure r
for the selected covariates when using \eqn{\alpha_{ER}}.}
\item{BestRE}{A vector of  relevance measure  r
for the selected covariates when using\eqn{\alpha_{RE}}.}
}
\description{
This function returns the selected covariates after estimating the optimal
threshold lambda to control FSR.
}
\examples{
lambdamax <- 10
lambda <- c(seq(1, lambdamax, length = 800))
data(FixCov)
data(VarCov)
data(counts)
print.progress = FALSE
m <- 5
option <- "OWN"
B <- 2
ncores <- 1
BBootBSOut<-csrnaseq:::BBootBS(B, ncores, counts[1:100,],
FixCov, VarCov, m,  lambda, option, print.progress)
DataBSOut <- csrnaseq:::DataBS(counts[1:100,], FixCov, VarCov,
lambda, print.progress)
FSRPBSOut <- csrnaseq:::FSRPBS(DataBSOut, BBootBSOut)
alpha0 <- 0.05
BestLambdaOut <- csrnaseq:::BestLambdaBS(FSRPBSOut, alpha0)
BestCovBSOut <- csrnaseq:::BestCovBS(DataBSOut, BestLambdaOut)
BestCovBSOut
}
\author{
Yet Nguyen \email{tienyettoan@gmail.com}.
}
