% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2-DataBS.R
\name{BootBS}
\alias{BootBS}
\title{Backward Selection for the Covariates Including Pseudo-Variables}
\usage{
BootBS(
  counts,
  FixCov,
  VarCov,
  m = 3,
  nrep,
  lambda,
  option = "OWN",
  print.progress = T
)
}
\arguments{
\item{counts}{a numerical  \code{matrix} of raw counts
Counts must be non-negative and NAs are not permitted.}

\item{FixCov}{a \code{data.frame} of the covariates that
are always included in the model.}

\item{VarCov}{a \code{data.frame} of the covariates subjected to  selection.}

\item{m}{number of generated pseudo-variables}

\item{nrep}{replication index for each pseudo-variables set.}

\item{lambda}{a vector of critical thresholds for the relevance measure.
The typical \code{lambda} has range from 1 to 10.}

\item{option}{the method to generate pseudo-variables
This is either one of these \code{"WN", "RX", "OWN", "ORX"}.}

\item{print.progress}{logical. If \code{TRUE} the print out the progress of
backward selection.}
}
\value{
a vector of 2*length(\code{lambda}) elements.
Its first length(\code{lambda})
elements are U(\eqn{\lambda}), the number of pseudo-variables selected at
the critical threshold \eqn{\lambda}.
Its last length(\eqn{\lambda}) elements are SP(\eqn{\lambda}), the total number
of selected covariates including pseudo-variables  at the critical threshold \eqn{\lambda}.
}
\description{
This function performs backward selection on the covariates
including available measured covariates and pseudo-variables and returns
a vector of the number of selected pseudo-variables U  and a vector
of the total number of selected covariates SP for the critical
threshold \code{lambda}.
}
\examples{
lambdamax <- 10
lambda <- c(seq(1, lambdamax, length = 800))
data(FixCov)
data(VarCov)
data(counts)
print.progress = FALSE
m <- 5
nrep <- 1
option <- "OWN"
BootBSOut <- csrnaseq:::BootBS(counts[1:100,], FixCov, VarCov, m,  nrep, lambda, option, print.progress)
length(BootBSOut)
head(BootBSOut)
}
\author{
Yet Nguyen \email{tienyettoan@gmail.com}.
}
