% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-data.R
\docType{data}
\name{counts}
\alias{counts}
\title{RFI RNA-seq Data}
\format{
The dataset has 12280 rows and 31 columns corresponding
to 31 RNA-seq samples from 31 pigs.
}
\usage{
counts
}
\description{
This is the RFI RNA-seq count data used  as an example for this study.
}
\examples{
data(counts)
dim(counts)
head(counts)
}
\keyword{datasets}
