% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-FSRBSSimulation.R
\name{DEAeval}
\alias{DEAeval}
\title{Calculate evaluation metrics on differential expression analysis}
\usage{
DEAeval(p, lab, method)
}
\arguments{
\item{p}{vector of p-values}

\item{lab}{vector of 0 and 1. 0 means true EE, 1 means true DE}

\item{method}{the methods that handles available covariates
(WN_RE, WN_ER, RX_ER, RX_RE, OWN_RE, OWN_ER, ORX_ER, ORX_RE, Oracle, OnlyLine, All, OldBS)}
}
\value{
A vector of 5 elements: NTP, R, V, FDP, PAUC
}
\description{
This function calculate the number of true positive (NTP),
the number of declared differentially expressed genes (R),
number of false positive (V), False Discovery Proportion (FDP),
and the partial area under ROC
curve with false positive rate less than 0.05 (PAUC) from the p-values vector
of the main factor of interest when analyzing simulated count data.
}
\examples{
sim <- readRDS("./analysis/SimulationOut/ModelSize_6_nGene_2000_B_100_m_7_alpha0_0.05/nrep_2.rds")
AllCov <- cbind(sim$SimCntOut$FixCov, sim$SimCntOut$VarCov[,sim$SimSelCov.ER$ORX.ER])
counts <- sim$SimCnt$SimCnt
p <- VoomPv(counts, AllCov)$pvs[,"Line"]
lab <-as.factor(as.numeric(!(sim$SimCnt$IndSample\%in\%sim$SimCnt$EEGene$Line)))
method <- "ORX.ER"
DEAevalOut <- DEAeval(p, lab, method)
DEAeval
}
