% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-multtest.R
\name{estimate.lambda}
\alias{estimate.lambda}
\title{Estimating The threshold lambda in Estimating FDR Using Storey's Formula}
\usage{
estimate.lambda(p, B = 20)
}
\arguments{
\item{p}{a numerical vector of p-value}

\item{B}{number of bin}
}
\value{
The function returns an estimate of lambdastar, the threshold
to calculate FDR based on Storey's formula.
}
\description{
This function calculate lambda threshold to estimate FDR
}
\examples{
p <- runif(100)
lambdastar <- estimate.lambda(p)
lambdastar
}
\references{
Dan NETTLETON, J. T. Gene HWANG, Rico A. CALDO, and Roger P. WISE.
Estimating the Number of True Null Hypotheses From a Histogram of p Values.
Journal of Agricultural, Biological, and Environmental Statistics,
Volume 11, Number 3, Pages 337–356.
}
\author{
Dan Nettleton \email{dnett@iastate.edu}
}
