% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-data.R
\docType{data}
\name{FixCov}
\alias{FixCov}
\title{Dataframe Containing The Set of Fix Covariates}
\format{
\code{FixCov} is a data frame containing only \code{Line}
}
\usage{
FixCov
}
\description{
This data set contains the covariates that are not subjected to variable
selection. In this case, it is \code{Line}, the main factor of
interest.
}
\examples{
data(FixCov)
names(FixCov)
FixCov
}
\keyword{datasets}
