% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-FSRProportion.R
\name{FSR}
\alias{FSR}
\title{Estimate False Selection Rate FSR}
\usage{
FSR(FSRPBSOut, BestLambdaOut)
}
\arguments{
\item{FSRPBSOut}{the output from \code{\link{FSRPBS}}.}

\item{BestLambdaOut}{the output from \code{\link{BestLambdaBS}}.}
}
\value{
A data frame with 3 columns
\item{lambda}{the numerical vector of the critical thresholds.}
\item{alphahat.RE}{estimate of \eqn{\alpha_{RE}}  at lambda.}
\item{alphahat.ER}{estimate of \eqn{\alpha_{ER}}  at lambda.}
}
\description{
This function estimates FSR  in two cases
\eqn{\alpha_{RE}} and \eqn{\alpha_{RE}}.
}
\examples{
lambdamax <- 10
lambda <- c(seq(1, lambdamax, length = 800))
data(FixCov)
data(VarCov)
data(counts)
print.progress = FALSE
m <- 5
option <- "OWN"
B <- 2
ncores <- 1
svamethod <- FALSE
BBootBSOut<-csrnaseq:::BBootBS(B, ncores, counts[1:100,],
FixCov, VarCov, m,  lambda, option, print.progress)
DataBSOut <- csrnaseq:::DataBS(counts[1:100,], FixCov, VarCov,
lambda, print.progress, svamethod)
FSRPBSOut <- csrnaseq:::FSRPBS(DataBSOut, BBootBSOut)
alpha0 <- 0.05
BestLambdaOut <- csrnaseq:::BestLambdaBS(FSRPBSOut, alpha0)
FSROut <- csrnaseq:::FSR(FSRPBSOut, BestLambdaOut)
dim(FSR)
head(FSR)
}
\author{
Yet Nguyen \email{tienyettoan@gmail.com}.
}
