% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-FSRProportion.R
\name{FSRAnalysisBS}
\alias{FSRAnalysisBS}
\title{Variable Selection in RNA-seq Differential Expression
Analysis Using Pseudo-Variables}
\usage{
FSRAnalysisBS(
  counts,
  FixCov,
  VarCov,
  option = "OWN",
  B = 100,
  m = 3,
  lambdamax = 10,
  alpha0 = 0.05,
  ncores,
  print.progress,
  saveall = TRUE
)
}
\arguments{
\item{counts}{a numerical  \code{matrix} of raw counts
Counts must be non-negative and NAs are not permitted.}

\item{FixCov}{a \code{data.frame} of the covariates that
are always included in the model.}

\item{VarCov}{a \code{data.frame} of the covariates subjected to  selection.}

\item{option}{the method to generate pseudo-variables
This is either one of these \code{"WN", "RX", "OWN", "ORX"}.}

\item{B}{a numerical value representing the number of
sets of the pseudo-variables. Usually, \code{B = 100} is
good enough for this method.}

\item{m}{number of generated pseudo-variables}

\item{lambdamax}{the maximum value of  lambda vector}

\item{alpha0}{the nominal FSR  level.}

\item{ncores}{number of cores to use for parallel computing via
\code{\link[parallel]{mclapply}}.}

\item{print.progress}{logical. If \code{TRUE} the print out the progress of
backward selection.}

\item{saveall}{logical. If it is TRUE, then save all output, otherwise, just save
the best covariate BestCovOut, which is the output from \link{BestCovBS}}
}
\value{
A list of the following components if \code{saveall = TRUE}
\item{DataBSOut}{ Output of DataBS function, see \link{DataBS}}
\item{BBootBSOut}{Output of BBootBS function, see \link{BBootBS}}
\item{FSROut}{Output of \link{FSR}}
\item{PlotFSROut}{Output plot of \link{PlotFSR}}
\item{BestLambdaOut}{Output of \link{BestLambdaBS}}
\item{BestCovOut}{Output of \link{BestCovBS}}
\item{VoomPvOutER}{Output of \link{PlotVoomPv} for ErrType ER}
\item{PlotVoomPvOutER}{Output plot of \link{PlotVoomPv} for ErrType ER}
\item{option}{Method generates pseudo-variables: WN, RX, OWN, ORX}
\item{B}{Number of sets of pseudo-variables}
\item{m}{Number of pseudo-variables}
}
\description{
This function performs the proposed backward variable selection
using pseudo-variables for RNA-seq differential expression analysis. The idea
is to select the most relevant covariates such that the false selection rate
is  below some pre-specified threshold.
}
\examples{
data(counts)
data(FixCov)
data(VarCov)
option <- "OWN"
B <- 5
m <- 5
lambdamax <- 5
alpha0 <- 0.05
ncores <- 1
print.progress <- FALSE
saveall <- TRUE
FSRAnalysisBSOut <- FSRAnalysisBS(counts, FixCov, VarCov, option, B, m, lambdamax, alpha0, ncores, print.progress, saveall)
names(FSRAnalysisBSOut)
}
