% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-FSRBSSimulation.R
\name{FSRnSimBS}
\alias{FSRnSimBS}
\title{Analyze multiple simulated datasets}
\usage{
FSRnSimBS(
  RealDataOut,
  nGene,
  nSim,
  B,
  m,
  lambdamax,
  alpha0,
  ncores,
  print.progress,
  saveall = FALSE,
  savesim = TRUE,
  foldername = "SimulationOut"
)
}
\arguments{
\item{RealDataOut}{the output of voom-limma differential expression analysis
implemented in \link{VoomPv} for RFI RNA-seq data. This analysis is performed in the
script 2-RSRBS_ModelSize_Simulation.R  in analysis folder.}

\item{nGene}{number of genes in simulated count data}

\item{nSim}{a vector of  n replication indices, in the form  1:n}

\item{B}{a numerical value representing the number of
sets of the pseudo-variables. Usually, \code{B = 100} is
good enough for this method.}

\item{m}{number of generated pseudo-variables}

\item{lambdamax}{the maximum value of  lambda vector}

\item{alpha0}{the nominal FSR  level.}

\item{ncores}{number of cores to use for parallel computing via
\code{\link[parallel]{mclapply}}.}

\item{print.progress}{logical. If \code{TRUE} the print out the progress of
backward selection.}

\item{saveall}{logical. If it is TRUE, then save all output, otherwise, just save
the best covariate BestCovOut, which is the output from \link{BestCovBS}}

\item{savesim}{logical. If TRUE, save all the output including SimCounts to
`SimulationOut/ModelSize_*_nGene_*_B_*_m_*_alpha0_*/nrep_*.rds`}

\item{foldername}{a character string that is the name of folder that saves
the result. Defaul is `foldername = "SimulationOut"  `.}
}
\value{
A list of 2 components
\item{TrueSelCov}{a vector of the name of the relevant covariates used in simulation}
\item{nSimSelCov}{a data frame of length(nSim) row, each coresponds to the set of FSP and FDP of the considered methods}
}
\description{
This function performs the analysis \link{FSR1SimBS} for length(nSim) simulated
datasets. For each nrep in nSim, it will call \link{FSR1SimBS}, then extract the number of
selected covariates (R), the number of selected relevant covariates  (S),
and the number of selected irrelevant covariates (U). Here, R = S + U.
Then it calculates the empirical False Selection Rate using either \eqn{\alpha_{RE}} or
\eqn{\alpha{ER}}.
}
\examples{
\dontrun{
RealDataOut <- readRDS("./analysis/RealDataOutBS/ModelSize_6.rds")
nGene <- 2000
nSim <- 1:3
B <- 4
m <- 2
lambdamax <- 5
alpha0 <- 0.05
ncores <- 4
print.progress <- FALSE
SimPath <- paste0("SimulationOut/ModelSize_", ModelSize, "_nGene_", nGene, "_B_", B, "_m_", m,  "_alpha0_", alpha0)
FSRnSimBSOut <- FSRnSimBS(RealDataOut, nGene, nSim, B, m, lambdamax, alpha0, ncores, print.progress, saveall = FALSE, savesim = TRUE)
}
}
