% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-FSRProportion.R
\name{FSRPBS}
\alias{FSRPBS}
\title{Estimate False Selection Rate Proportion}
\usage{
FSRPBS(DataBSOut, BBootBSOut)
}
\arguments{
\item{DataBSOut}{the output from function \code{\link{DataBS}}.}

\item{BBootBSOut}{the output from function \code{\link{BBootBS}}.}
}
\value{
A list of 6 components
\item{B}{a numerical value representing the number of
sets of the pseudo-variables.}
\item{nPhoCov}{the number of pseudo-covariates.}
\item{BS}{the \code{BS} component of the output DataBSOut from function
\code{\link{DataBS}}, which is a vector of \code{r} value of the eliminated covariates
during the selection process.}
\item{lambda}{the \code{lambda} component of the output DataBSOut from
the function \code{\link{DataBS}}, which is a numerical vector of the critical thresholds
\eqn{\lambda}.}
\item{alphahat.P.RE}{\eqn{\alpha_{RE,P}} defined as above. }
\item{alphahat.P.ER}{\eqn{\alpha_{ER,P}} defined as above. }
}
\description{
This function  estimates  the False Selection Rate Proportion (FSRP)
via the pseudo-variable. \eqn{\alpha_{RE,P}} is based
on Ratio of Expectation, \eqn{\alpha_{ER,P}} is based on
Expectation of Ratio.
}
\details{
\deqn{\alpha_{RE, P} = \frac{U}{1+SP}}
where \code{SP} is the average of the total number of selected covariates among both available covariates and
pseudo-variables, U is the average of the number of selected pseudo-variables over B replications.
\deqn{\alpha_{ER, P} = \frac{U}{1+S}}
where \code{S} is the number of covariates selected in the set
of the available covariates,  U is the average of the number of selected pseudo-variables over B replications.
}
\examples{
lambdamax <- 5
lambda <- c(seq(1, lambdamax, length = 400))
data(FixCov)
data(VarCov)
data(counts)
print.progress = FALSE
m <- 5
option <- "OWN"
B <- 2
ncores <- 1
svamethod <- FALSE
BBootBSOut<-csrnaseq:::BBootBS(B, ncores, counts[1:100,],
FixCov, VarCov, m,  lambda, option, print.progress)
DataBSOut <- csrnaseq:::DataBS(counts[1:100,], FixCov, VarCov,
lambda, print.progress, svamethod)
FSRPBSOut <- csrnaseq:::FSRPBS(DataBSOut, BBootBSOut)
names(FSRPBSOut)
}
\author{
Yet Nguyen \email{tienyettoan@gmail.com}
}
