% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-bs.R
\name{jabes.bs}
\alias{jabes.bs}
\title{Backward Selection Algorithm  in Nguyen et al. 2015, JABES}
\usage{
jabes.bs(counts, FixCov, VarCov, print.progress = FALSE)
}
\arguments{
\item{counts}{a numerical  \code{matrix} of raw counts
Counts must be non-negative and NAs are not permitted.}

\item{FixCov}{a \code{data.frame} of the covariates that
are always included in the model.}

\item{VarCov}{a \code{data.frame} of the covariates subjected to  selection.}

\item{print.progress}{logical. If \code{TRUE} the print out the progress of
backward selection.}
}
\value{
A list of 5 components
\item{WorstP5}{a vector of the number of p-values less than 0.05 for each
eliminated covariate during the selection process. The first element of the vector is the first
covariate eliminated, and so on.}
\item{Q5}{a vector of the number of q-values less than
0.05 for the test of the effect of the main variable of interest
at each iteration.}
\item{VarCov}{it is the same as the input argument.}
\item{FixCov}{it is the same as the input argument.}
\item{BestCov}{ the set of selected covariates from
the list of sequences of subset of covariates remaining
at each iteration of the backward selection. }
}
\description{
This function implements the backward selection algorithm
in Nguyen et al. 2015, JABES. The covariate relevance measure
 is the number of p-values less than
0.05. The algorithm is applied to the data until the main
factor of interest is eliminated from the model. Among the
 sets of remaining covariates, one that possesses the highest number
 of DDE genes (FDR <= 0.05) w.r.t. the main factor of interest
 is the final set of covariates selected by the algorithm.
}
\examples{
data(counts)
data(FixCov)
data(VarCov)
print.progress <- FALSE
jabes.bsout <- jabes.bs(counts, FixCov, VarCov, print.progress)
names(jabes.bsout)
jabes.bsout$Q5
jabes.bsout$BestCov
}
\references{
Yet Nguyen, Dan Nettleton, Haibo Liu, Christopher K. Tuggle.
Detecting Differentially Expressed Genes with
RNA-seq Data Using Backward Selection to Account
for the Effects of Relevant Covariates.
J Agric Biol Environ Stat. 2015; 20(4): 577–597.
}
\author{
Yet Nguyen \email{tienyettoan@gmail.com}
}
