% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-multtest.R
\name{jabes.q}
\alias{jabes.q}
\title{Q-value Using Histogram-based Method (Nettleton 2006)}
\usage{
jabes.q(p, B = 20)
}
\arguments{
\item{p}{a numerical vector of p-value}

\item{B}{number of bin}
}
\value{
The function returns a q-value vector of the input p-value vector.
}
\description{
This function computes q-values using the approach of Nettleton et al.
(2006) JABES 11, 337-356.  Author: Dan Nettleton
}
\examples{
p <- runif(100)
q <- jabes.q(p)
sum(q <= .05)
}
\references{
Dan NETTLETON, J. T. Gene HWANG, Rico A. CALDO, and Roger P. WISE.
Estimating the Number of True Null Hypotheses From a Histogram of p Values.
Journal of Agricultural, Biological, and Environmental Statistics,
Volume 11, Number 3, Pages 337–356.
}
\author{
Dan Nettleton \email{dnett@iastate.edu}
}
