% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1-bs.R
\name{kss}
\alias{kss}
\title{Function Calculates KSG distance}
\usage{
kss(z)
}
\arguments{
\item{z}{a vector of p-values}
}
\value{
Kolmogorov-Smirnov distance of Grenander estimator of p-values's ECDF
and uniform distribution u(0,1)
}
\description{
Description: This function calculates Komogorov-Smirnov Grenander Distance of
the empirical cummulative distribution function of the p-value vector and uniform unif(0, 1)
}
