% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-FSRProportion.R
\name{PlotVoomPv}
\alias{PlotVoomPv}
\title{Histogram of p-values of covariates}
\usage{
PlotVoomPv(
  VoomPvOut,
  option = "OWN",
  B = 100,
  m = 3,
  ErrType = "ER",
  alpha0 = 0.05
)
}
\arguments{
\item{VoomPvOut}{The output of \code{\link{VoomPv}}}

\item{option}{the method to generate pseudo-variables
This is either one of these \code{"WN", "RX", "OWN", "ORX"}.}

\item{B}{a numerical value representing the number of
sets of the pseudo-variables. Usually, \code{B = 100} is
good enough for this method.}

\item{m}{number of generated pseudo-variables}

\item{ErrType}{either "ER" (for \eqn{\alpha_{ER}})  or "RE" (for \eqn{\alpha_{RE}})}

\item{alpha0}{the nominal FSR  level.}

\item{lambdamax}{the maximum value of  lambda vector}
}
\value{
A plot of histogram of p-values for the covariates of the final model
}
\description{
This function plots a histogram of the p-values for all covariates,
the primary ones and those subjected to variable selection.
}
\examples{
lambdamax <- 5
lambda <- c(seq(1, lambdamax, length = 400))
data(FixCov)
data(VarCov)
data(counts)
print.progress = FALSE
m <- 5
option <- "OWN"
B <- 5
ncores <- 1
alpha0 <- 0.05
DataBSOut <- DataBS(counts = counts[1:1000,],
FixCov = FixCov, VarCov=VarCov, lambda = lambda,
print.progress = print.progress, svamethod = svamethod)
BBootBSOut <- BBootBS(B= B, ncores = ncores, counts = counts[1:1000,],
FixCov, VarCov = DataBSOut$VarCov, m= m, lambda = lambda,
option = option, print.progress = print.progress)
FSRPBSOut <- FSRPBS(DataBSOut=DataBSOut, BBootBSOut = BBootBSOut)
BestLambdaOut <- BestLambdaBS(FSRPBSOut, alpha0)
FSROut <- FSR(FSRPBSOut, BestLambdaOut)
PlotFSROut <- PlotFSR(FSROut = FSROut, alpha0 = alpha0, option = option, B = B, m = m)
BestCovOut <- BestCovBS(DataBSOut = DataBSOut, BestLambdaOut = BestLambdaOut)
VoomPvOutER <- VoomPv(counts = counts,
                      AllCov = cbind(FixCov, (DataBSOut$VarCov)[names(BestCovOut$BestER)]))
PlotVoomPvOutER <-  PlotVoomPv(VoomPvOut = VoomPvOutER,
option = option, B = B, m = m, ErrType = "ER", alpha0 = alpha0)
PlotVoomPvOutER
VoomPvOutRE <- VoomPv(counts = counts,
AllCov = cbind(FixCov, (DataBSOut$VarCov)[names(BestCovOut$BestRE)]))
PlotVoomPvOutRE <-  PlotVoomPv(VoomPvOut = VoomPvOutRE,
option = option, B = B, m = m, ErrType = "RE", alpha0 = alpha0)
PlotVoomPvOutRE
}
