% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/4-FSRBSSimulation.R
\name{SimCounts}
\alias{SimCounts}
\title{Simulate Count Data from RFI RNA-seq Data}
\usage{
SimCounts(RealDataOut, nGene = 2000, nrep)
}
\arguments{
\item{RealDataOut}{the output of voom-limma differential expression analysis
implemented in \link{VoomPv} for RFI RNA-seq data. This analysis is performed in the
script 2-RSRBS_ModelSize_Simulation.R  in analysis folder.}

\item{nGene}{number of genes in simulated count data}

\item{nrep}{replication index, used in set.seed within this function}
}
\value{
A list of 3 components
\item{SimCnt}{simulated count data of nGene rows and 31 columns}
\item{IndSample}{a vector of original row ID of the RFI RNA-seq data for nGene simulated genes}
\item{EEGene}{a list of vectors of original row ID of the RFI RNA-seq data for  null genes of each covariate}
}
\description{
This function simulates count datasets that contain both EE and DE genes
  with respect to the
 primary variable and each of the relevant covariates. For each simulation scenario,
 as true parameters to simulate new data, we used the precision weights,
 the scaled error variances and the partial regression coefficient estimates
 from the `voom-limma` fit \link{VoomPV} of the corresponding model to the RFI RNA-seq data,
 except that we set partial regression coefficients on each variable to zero
 for a subset of genes to allow simulation of EE genes. More specifically,
 for each variable \eqn{j} (either relevant covariates or the primary variable
 Line), the \eqn{\hat{G}_0^{(j)}} least significant partial regression
coefficients were set to zero, where  \eqn{\hat{G}_0^{(j)}}  is the estimated
number of the \eqn{j}-covariate partial regression coefficients equal to zero
when the histogram-based method of Nettleton et al. 2006 is applied to the
\eqn{j}-variable's \eqn{p}-values
from the \link{VoomPv} fit of the corresponding model to the RFI RNA-seq data.
This strategy yielded a parameter vector (consisting of a scaled error
variance, precision weights, and  partial regression coefficients) for
each of 12280 genes. To simulate any particular dataset for a given set
of truly relevant covariates (either 0, 1, 2, 6, 7, or 8 relevant covariates),
we randomly sampled \code{nGene} (default value is 2000) gene parameter vectors.
The selected parameters
and the respective  values of the primary variable and the relevant
covariates for the 31 pigs were used to simulate a \code{nGene} x 31 dataset
of read counts following the inverse steps of log2 count transformation and
the linear model used in the main paper.
}
\examples{
\dontrun{
RealDataOut <- readRDS("./analysis/RealDataOutBS/ModelSize_Simulation_6.rds")
nGene <- 2000
nrep <- 1
SimCnt <- SimCounts(RealDataOut, nGene,nrep)
}
}
\author{
Yet Nguyen \email{tienyettoan@gmail.com}
}
