% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0-data.R
\docType{data}
\name{VarCov}
\alias{VarCov}
\title{A Data frame of Covariates}
\format{
\code{VarCov} is a data frame containing 13 covariates.
}
\usage{
VarCov
}
\description{
This data set contains the covariates subjected to variable
selection. In this case, it includes \code{RFI, Diet, Neut, Lymp, Mono,
Neut, Baso, Eosi, Conca, Concb, RINa, RINb, Block, Order.}
}
\examples{
data(VarCov)
names(VarCov)
VarCov
}
\keyword{datasets}
