% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3-FSRProportion.R
\name{VoomPv}
\alias{VoomPv}
\title{Differential expression analysis using voom-limma}
\usage{
VoomPv(counts, AllCov)
}
\arguments{
\item{counts}{an RNA-seq read count data}

\item{AllCov}{a data frame of all covariates}
}
\value{
A list of the following elements
\item{y}{ A matrix of logcpm count data}
\item{pvs}{A matrix of p-value vectors, one p-value vector for each covariate}
\item{Beta}{A matrix of regression coefficent estimates}
\item{Yhat}{A matrix of fitted values}
\item{lib.size}{A vector of ibrary size  (75quartile)}
\item{weights}{A matrix of observational weights}
\item{sigma}{sigma from output of \code{\link[limma]{lmFit}}}.
\item{design}{the design matrix}
}
\description{
This function performs voom limma differential expressin analysis and returns  selected
outputs of the \code{\link[limma]{voom}} result including a
vector of p-values, transformed counts,...
}
\examples{
data(FixCov)
data(VarCov)
data(counts)
AllCov <- cbind(FixCov, VarCov)
VoomPvOut <- csrnaseq:::VoomPv(counts, AllCov)
names(VoomPvOut)
}
